prototype search mixin
